Split bbcode.blocks into separate modules
