+++ misago/markup/bbcode/quote.py
@@ -2,20 +2,12 @@ import re
 
 import markdown
 from django.utils.crypto import get_random_string
-from markdown.blockprocessors import BlockProcessor, HRProcessor
-from markdown.extensions.fenced_code import FencedBlockPreprocessor
+from markdown.blockprocessors import BlockProcessor
 from markdown.preprocessors import Preprocessor
 from markdown.util import etree
 
 QUOTE_START = get_random_string(32)
 QUOTE_END = get_random_string(32)
-
-
-class BBCodeHRProcessor(HRProcessor):
-    RE = r"^\[hr\]*"
-
-    # Detect hr on any line of a block.
-    SEARCH_RE = re.compile(RE, re.MULTILINE | re.IGNORECASE)
 
 
 class QuoteExtension(markdown.Extension):
@@ -102,21 +94,3 @@ class QuoteBlockProcessor(BlockProcessor):
                 heading.text = title
 
             self.parser.parseBlocks(blockquote, children)
-
-
-class CodeBlockExtension(markdown.Extension):
-    def extendMarkdown(self, md):
-        md.registerExtension(self)
-
-        md.preprocessors.add(
-            "misago_code_bbcode", CodeBlockPreprocessor(md), ">normalize_whitespace"
-        )
-
-
-class CodeBlockPreprocessor(FencedBlockPreprocessor):
-    FENCED_BLOCK_RE = re.compile(
-        r"""
-\[code(=("?)(?P<lang>.*?)("?))?](([ ]*\n)+)?(?P<code>.*?)((\s|\n)+)?\[/code\]
-""",
-        re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE,
-    )
